refactor HTTP module
