+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "colyseus.js",
-  "version": "0.15.14",
+  "version": "0.15.15-alpha.0",
   "description": "Colyseus Multiplayer SDK for JavaScript/TypeScript",
   "author": "Endel Dreyer",
   "license": "MIT",
